Add files via upload
Hey guys, this is the generic user profile page that I created. I didn't follow the color scheme or form of the previous web page since this page will be temporary and I will edit it later on. I've included the aspects of the username, profile picture as well as skills as requested. Can you guys review it and give me your opinion?
Thank you,
Mario Bastiampillai