Additional changes to the CSS of the Profile Page #138
The background images were changed once again to fit more with the style present throughout the rest of the site to make it more uniform. A thick green border was added to encapsulate the two different containers so that they may appear more presentable. The font present in the credentials area was changed once again.